Codigos de errores para delete y get NotaMaterias

+++ Servicio/servicio.py
@@ -3,9 +3,9 @@ from flask import jsonify
 import json
 from controller import Response
 from flask_api import FlaskAPI, status
-from Servicio.Exception_api import ApiExceptionServ,NotFound,InternalServerError
+from Servicio.Exception_api import *
 import requests
-from Servicio.Errores_internos import CodeInternalError
+from Servicio.Internal_errors import CodeInternalError
 
 
 #add notaMateria
@@ -22,15 +22,17 @@ def addNotaMateria(request):
     '''return "Materia agregarda. id={}".format(notamateria.notamateria_id)'''
 
 #get
-def findNotasMateriasByAlumnoID(id):                 
-        obj=NotaMateria.getNotasMateriasByAlumnoID(id) 
-        if obj is not None and len(obj)!=0:
-            json_Str=jsonify([e.serializar() for e in obj]) 
-            return json_Str
-        else:
-            raise NotFound()
-            #content = {'detalle': 'Recurso no encontrado.','cod_error_interno':'2','descripcion_error':'No se encontraron registros con los datos ingresados.'}
-            #return content, status.HTTP_404_NOT_FOUND,content'''
+def findNotasMateriasByAlumnoID(id):
+    try:
+        obj=NotaMateria.getNotasMateriasByAlumnoID(id)
+    except Exception as identifier:
+        raise InternalServerError('Error relacionado con base de datos.', CodeInternalError.ERROR_INTERNO_11_CONEXION_BD)          
+    if obj is not None and len(obj)!=0:
+        json_Str=jsonify([e.serializar() for e in obj]) 
+        return json_Str
+    else:
+        raise BadResquest('Recurso no encontrado en la base de datos.',CodeInternalError.ERROR_INTERNO_12_REQUEST_NOT_FOUND)
+            
 
 # Get all NotasMaterias
 def getNotasMaterias():
@@ -69,18 +71,19 @@ def updateNotaMateriasByAlumnoID(id):
 
 #
 def deleteNotaMateria(id):
-    notamateria=NotaMateria.buscarNotaMateriaByNotamateriaID(id)
-    notamateria.delete()
-    return ('Registro eliminado') 
-    '''try:
-        1==1
-    except Exception as e:
-        return TaskNotFound(e)
-    finally
-        return TaskNotFound(APIException)'''
-    #notamateria=NotaMateria.buscarNotaMateriaByNotamateriaID(id)
-    #notamateria.delete()
-    #return ('se elimino')  
+    try:
+        notamateria=NotaMateria.buscarNotaMateriaByNotamateriaID(id)
+    except Exception as identifier:
+        raise InternalServerError('Error relacionado con base de datos.', CodeInternalError.ERROR_INTERNO_11_CONEXION_BD)          
+    if notamateria is not None:
+        try:
+            notamateria.delete()
+            return ('Recurso eliminado.',status.HTTP_200_OK)
+        except Exception as identifier:
+            raise InternalServerError('Error relacionado con base de datos.', CodeInternalError.ERROR_INTERNO_11_CONEXION_BD)   
+    else:
+        raise BadResquest('Recurso no encontrado en la base de datos.',CodeInternalError.ERROR_INTERNO_12_REQUEST_NOT_FOUND)  
+     
 
 #Methods
 def setNotaMateria(request):
@@ -91,13 +94,13 @@ def setNotaMateria(request):
             request.json['notafinal'] 
         )
     except Exception as identifier:
-        raise NotFound('Estructa de Json incorrecta',CodeInternalError.ERROR_INTERNO_10_JSON_BAD_FORMED)
+        raise BadResquest('Estructa de Json incorrecta',CodeInternalError.ERROR_INTERNO_10_JSON_BAD_FORMED)
     return notamateria
        
 
        
-def setMessajeFormatJson():
-    return ({'detail':'Estructura json no soportada'},status.HTTP_400_BAD_REQUEST)
+#def setMessajeFormatJson():
+#    return ({'detail':'Estructura json no soportada'},status.HTTP_400_BAD_REQUEST)
 #
 
 def imprimirJson():
